% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_results.R
\name{create_results}
\alias{create_results}
\title{Create results}
\usage{
create_results(area, max_increase, rule_distance_value_m, sites)
}
\arguments{
\item{area}{Vector of area within each isoline}

\item{max_increase}{Level where the area increase has its (local) maximum. Can either be set manually or automatically by using ?TODO?}

\item{rule_distance_value_m}{Rule-distance in metres}

\item{sites}{SpatialPointsDataFrame of known sites}
}
\value{
A dataframe containing the results, such as "Total area in square kilometres", "Expected rule-distance in  kilometers", "Hexagon area in square kilometres", "Number of expected sites", and "Percentage of known sites"
}
\description{
This function produces the final results
}
\author{
David N. Matzig
}
