% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rule_distance.R
\name{density_plot_nn}
\alias{density_plot_nn}
\title{Determine the rule-distance}
\usage{
density_plot_nn(project_CRS, sites_table, xy, threshold = 10000,
  bin_width = 250, x_axis_steps = 1000)
}
\arguments{
\item{project_CRS}{Coordinate Reference System of Project. Unit has to be metres.}

\item{sites_table}{Dataframe or tibble containing individual sites in each row with two separate columns containing the coordinates}

\item{xy}{Dataframe or tibble containing the coordinates in two separate columns}

\item{threshold}{Threshold for nearest neighbour in metres. Default is 10000}

\item{bin_width}{Bin width for density plot. Default is 250}

\item{x_axis_steps}{Steps on x-axis in metres. Default is 500}
}
\description{
This function helps finding the rule-distance. It produces a SpatialPointsDataframe, calculates the nearest neighbour distance and plots a density curve for the to determine the rule-distance.
}
\examples{

# provide CRS and data

project_CRS <- "+init=epsg:32634"
sites_table <- bronze_age_fortifications
xy <- sites_table[,c("xUTM","yUTM")]


# create density plot

density_plot_nn(project_CRS, sites, xy, threshold, bin_width, x_axis_steps)


# save rule-distance in variable 'rule_distance_value_m'

rule_distance_value_m <- 2500 
}
\author{
David N. Matzig
}
