% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largest_empty_circle.R
\name{largest_empty_circle}
\alias{largest_empty_circle}
\title{Calculate the largest empty circle raster}
\usage{
largest_empty_circle(sites_sf, raster = "default")
}
\arguments{
\item{sites_sf}{sf feature of the locations of interest}

\item{raster}{raster or SpatialGridDataFrame indicating the interpolation environment}
}
\value{
a list containing the LEC_raster and the Nodes including the distance to the closest location
}
\description{
This function calculates the interpolated largest empty circle raster for point distributions
}
\examples{
library(mapview)
library(sf)

locations <- bronze_age_fortifications \%>\% 
  st_as_sf(coords = c("xUTM", "yUTM"),crs = 32634)

applylarge <- largest_empty_circle(locations)
mapview(applylarge$LEC_raster)+mapview(applylarge$Nodes,zcol="mindist")+mapview(locations)

applylarge_truezero <- largest_empty_circle(locations,truezero=TRUE)
mapview(applylarge_truezero$LEC_raster)+mapview(applylarge_truezero$Nodes,zcol="mindist")+mapview(locations)
}
\author{
Wolfgang Hamer

Daniel Knitter
}
