% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_area_increase.R
\name{max_area_increase}
\alias{max_area_increase}
\title{Determine the point of maximum area increase}
\usage{
max_area_increase(project_CRS, sites_table, xy, x_axis_steps = 10)
}
\arguments{
\item{project_CRS}{Coordinate Reference System of Project. Unit has to be metres.}

\item{sites_table}{Dataframe or tibble containing individual sites in each row with two separate columns containing the coordinates}

\item{xy}{Dataframe or tibble containing the coordinates in two separate columns}

\item{x_axis_steps}{Steps on x-axis in metres. Default is 10}
}
\description{
This function helps finding the point of maximum area increase
}
\examples{

# provide CRS and data

project_CRS <- "+init=epsg:32634"
sites_table <- bronze_age_fortifications
xy <- sites_table[,c("xUTM","yUTM")]

# create density plot

density_plot_nn(project_CRS, sites, xy, threshold, bin_width, x_axis_steps)

# Plot ratio of area increase 

max_area_increase(project_CRS, sites_table, xy)


# set value for level of maximum area increase

max_increase <- 10
}
\author{
David N. Matzig
}
