% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecAAR-produce_results.R
\name{create_results}
\alias{create_results}
\title{Create results}
\usage{
create_results(max_increase, rule_distance_value_m)
}
\arguments{
\item{max_increase}{Level where the area increase has its (local) maximum. Can either be set manually or automatically by using ?TODO?}

\item{rule_distance_value_m}{Rule-distance in metres}
}
\value{
A dataframe containing the results, such as "Total area in square kilometres", "Expected rule-distance in  kilometers", "Hexagon area in square kilometres", "Number of expected sites", and "Percentage of known sites"
}
\description{
This function produces the final results
}
\examples{


# provide CRS and data

project_CRS <- "+init=epsg:32634"
sites_table <- bronze_age_fortifications
xy <- sites_table[,c("xUTM","yUTM")]


# 1. Create density plot

density_plot_nn(project_CRS, sites_table, xy)


# 2. look up the rule-distance from the plot and save it in the variable 'rule_distance_value_m'

rule_distance_value_m <- 2500 


# 3. Plot ratio of area increase 

max_area_increase(project_CRS, sites_table, xy)


# 4. Set the value for level of maximum area increase

max_increase <- 10


# Create table containing the results, such as "Total area in square kilometres", "Expected rule-distance in  kilometers", "Hexagon area in square kilometres", "Number of expected sites", and "Percentage of known sites"

create_results(max_increase, rule_distance_value_m)
}
\author{
David N. Matzig
}
